compatibility layer for grid_sample & failed UT fixes

+++ tests/test_output.py
@@ -148,21 +148,21 @@ class TestOutput(TestCase):
                 [[[1, 2]]],  # make sure it works for N-d features
                 [[[1, 3]]],
                 [[[1, 4]]],
-            ])
+            ], dtype=torch.float32)
 
         samples_p = torch.tensor(
             [
                 [[[1, 2.1]]],
                 [[[1, 3.1]]],
                 [[[1, 4.1]]],
-            ])
+            ], dtype=torch.float32)
 
         samples_n = torch.tensor(
             [
                 [[[1, 20.1]]],  # far from the samples_p and margin, should be 0 loss
                 [[[1, 3.1]]],  # not satisfying the margin
                 [[[1, 40.1]]],  # far from the samples_p and margin, should be 0 loss
-            ])
+            ], dtype=torch.float32)
 
         weights = torch.tensor([1.0, 3.0, 1.0])
         o = trw.train.OutputTriplets(samples, samples_p, samples_n, weight_name='weights')
